projects/release-worker/log-worker/lib/requirements.txt
xlwings==0.27.15
progressbar2==3.55.0

# test
pytest==8.3.2
